Revert "Swap function refactoring for royale"
This reverts commit 2d112a7c6349de7f974fd6e7146e360b99b1d2f8.
