+++ src/pages/Options/Royal/V2/components/RoyaleArena.tsx
@@ -31,8 +31,6 @@ type RoyaleArenaProps = {
     latestSeason: number;
     selectedSeason: number;
     showBattle: boolean;
-    buyInToken: any;
-    updateBalanceAndAllowance: (token: any) => void;
 };
 
 const renderRounds = (
@@ -222,8 +220,6 @@ export const RoyaleArena: React.FC<RoyaleArenaProps> = ({
     ethPrice,
     positions,
     royaleFooterData,
-    buyInToken,
-    updateBalanceAndAllowance,
 }) => {
     const { t } = useTranslation();
     const isAppReady = useSelector((state: RootState) => getIsAppReady(state));
@@ -314,7 +310,6 @@ export const RoyaleArena: React.FC<RoyaleArenaProps> = ({
 
             if (txResult && txResult.events) {
                 dispatchMarketNotification('Reward claimed');
-                updateBalanceAndAllowance(buyInToken);
             }
         }
     };
